Refactor CLI to Pure Projections (Hollow Data Plane)
- Removed all kinetic CLI entry points from `pyproject.toml`.
- Salvaged pure data transformations from `src/coreason_manifest/cli/` into `src/coreason_manifest/utils/projections.py`.
- Completely deleted the `src/coreason_manifest/cli/` directory.
- Updated export boundaries in `src/coreason_manifest/__init__.py`.
- Refactored tests in `tests/contracts/test_projections.py` to directly call pure functions without mocking sys.argv/sys.exit.
- Passed all linter, formatting, type checking, and testing checks.
