+++ src/coreason_manifest/__init__.py
@@ -8,6 +8,12 @@ from coreason_manifest.utils.algebra import (
     generate_correction_prompt,
     verify_merkle_proof,
 )
+from coreason_manifest.utils.projections import (
+    get_ontology_schema,
+    project_manifest_to_markdown,
+    project_manifest_to_mermaid,
+    validate_payload,
+)
 
 __all__ = [
     "ActionSpaceManifest",
@@ -281,5 +287,9 @@ __all__ = [
     "align_semantic_manifolds",
     "compute_topology_hash",
     "generate_correction_prompt",
+    "get_ontology_schema",
+    "project_manifest_to_markdown",
+    "project_manifest_to_mermaid",
+    "validate_payload",
     "verify_merkle_proof",
 ]
